services/vocoder/requirements.txt
grpcio==1.48.2
grpcio-tools==1.48.2
protobuf>=3.19.0,<3.20
numpy>=1.24.0

# Vocoder model (install when ready to use real HiFi-GAN)
# torch>=2.0.0
